Modified transducer codegen to write tranducer rep. to buffer rather than outstream.
The function fsm_transducer is responsible for converting an internal representation of a transducer and associated data structures into source code. That is, it is responsible for transducer code generation. Previously, the code was written directly to an output stream provided as an argument to the function.  However, we found this design to be too restrictive as it forces an all-or-nothing style of codegen: all the bindings generated by fsm_transducer appear together in the output, which gives the calling function little flexibility with their arrangement.

This commit changes that structure somewhat.  Now, fsm_transducer splits codegen into two parts. All "helper" bindings are still written directly to the output stream, as before. However, the representation of the transducer itself is written instead to an output buffer which is returned by the function. The caller can then print this representation when and where it pleases.