Instated single-pass codegen (for arrow) as the default.
I've modified the defaults in Compileopt to generate transducer code in a single pass. This required flipping a number of options. Since these options are related, I also added a general consistency-checker function to Compileopt that ensures the set of options are consistent.  I've modified Main to call this consistentcy checker.